+++ pkg/sim/zone_field.go
@@ -15,7 +15,7 @@ func (f Field) GetTiles() []TilePosition {
 	return f.Tiles
 }
 
-func (sim *Sim) CreateField(tiles []TilePosition, seedVariant int) *Field {
+func (sim *Sim) CreateField(tiles []TilePosition, seedVariant int16) *Field {
 	newField := Field{
 		Centroid:    GetZoneCentroid(tiles),
 		Tiles:       tiles,
@@ -119,7 +119,7 @@ func (sim *Sim) GetSuitableFieldTiles(character *Character) []TilePosition {
 
 				// Check all neighbors
 				for _, dir := range EightDirections {
-					newX, newY := current.X+dir[0], current.Y+dir[1]
+					newX, newY := current.X+int16(dir[0]), current.Y+int16(dir[1])
 
 					// Check bounds
 					if newX < 0 || newX >= config.RegionSize || newY < 0 || newY >= config.RegionSize {
